fix: change mesh typing for vertex in input

+++ headers/utility/graphics/mesh.hpp
@@ -12,17 +12,17 @@
 namespace utility::graphics {
     class Mesh {
     public:
-        Mesh(const std::vector<Vertex<float, float>> &vertices, const std::vector<uint32_t> &indices);
+        Mesh(const std::vector<Vertex<float, uint8_t>> &vertices, const std::vector<uint32_t> &indices);
         ~Mesh();
 
-        void addVertex(const Vertex<float, float> &vertex);
+        void addVertex(const Vertex<float, uint8_t> &vertex);
         void addIndex(uint32_t index);
 
-        const std::vector<Vertex<float, float>> &getVertices() const;
+        const std::vector<Vertex<float, uint8_t>> &getVertices() const;
         const std::vector<uint32_t> &getIndices() const;
 
     private:
-        std::vector<Vertex<float, float>> _vertices;
+        std::vector<Vertex<float, uint8_t>> _vertices;
         std::vector<uint32_t> _indices;
     };
 } // namespace utility::graphics
